python/requirements.txt
celery==4.3.0
Flask
redis
Werkzeug
python-dotenv
vine==1.3.0